requirements.txt
amqp==2.2.1
azure==2.0.0
billiard==3.5.0.3
celery==4.1.0
Django==1.11.4
django-extensions==1.8.1
django-filter==1.0.4
django-filters==0.2.1
django-polymorphic==1.3
django-simple-history==1.9.0
django-taggit==0.22.1
django-taggit-serializer==0.1.5
djangorestframework==3.6.3
kombu==4.1.0
pandas==0.20.3
paramiko==2.2.1
psycopg2==2.7.3
pytz==2017.2
six==1.10.0
supervisor==3.3.3
tables==3.4.2
vine==1.1.4
xlrd==1.1.0
